+++ src/openeo_aggregator/backend.py
@@ -663,15 +663,34 @@ class AggregatorBackendImplementation(OpenEoBackendImplementation):
         response.status_code = overall_status_code
         return response
 
+    BILLING_PLAN_FREE = "free"
+    BILLING_PLAN_EARLY_ADOPTER = "early-adopter"
+    BILLING_PLAN_VSC4 = "vsc4"
+
     def capabilities_billing(self) -> dict:
+        # TODO: ok to hardcode this here, or move to config?
         return {
             "currency": "EUR",
             "plans": [
                 {
-                    "name": "early-adopter",
+                    "name": self.BILLING_PLAN_EARLY_ADOPTER,
                     "description": "openEO.cloud early adopter plan",
                     "url": "https://openeo.cloud/early-adopters/",
-                    "paid": True
+                    "paid": True,
+                },
+                # TODO: Unused plan at the moment: necessary to expose it?
+                {
+                    "name": self.BILLING_PLAN_FREE,
+                    "description": "openEO.cloud free plan (experimental)",
+                    # TODO: url?
+                    "paid": False
+                },
+                # TODO: Unused plan at the moment: necessary to expose it?
+                {
+                    "name": self.BILLING_PLAN_VSC4,
+                    "description": "[EODC] Use this plan to process data on VSC-4",
+                    # TODO: url?
+                    "paid": True,
                 }
             ]
         }