+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- expose openEO Platform billing plans ([6](https://github.com/Open-EO/openeo-aggregator/issues/6))
+
 ### Changed
 
 - Refresh back-end connection objects regularly instead of holding on the same ones.